+++ contracts/RiscVInstructions/AtomicInstructions.sol
@@ -7,7 +7,7 @@ import "../../contracts/VirtualMemory.sol";
 
 library AtomicInstructions {
 
-  function execute_LR(MemoryInteractor mi, uint256 mmIndex, uint64 pc, uint32 insn, uint256 wordSize)
+  function execute_LR(MemoryInteractor mi, uint256 mmIndex, uint64 pc, uint32 insn, uint64 wordSize)
   public returns (bool) {
     uint64 vaddr = mi.read_x(mmIndex, RiscVDecoder.insn_rs1(insn));
     (bool succ, uint64 val) = VirtualMemory.read_virtual_memory(mi, mmIndex, wordSize, vaddr);
@@ -48,7 +48,7 @@ library AtomicInstructions {
     return true;
   }
 
-  function execute_AMO_part1(MemoryInteractor mi, uint256 mmIndex, uint64 pc, uint32 insn, uint256 wordSize)
+  function execute_AMO_part1(MemoryInteractor mi, uint256 mmIndex, uint64 pc, uint32 insn, uint64 wordSize)
   internal returns (uint64, uint64, uint64, bool){
     uint64 vaddr = mi.read_x(mmIndex, RiscVDecoder.insn_rs1(insn));
